refactor(fbuild-deploy): split esp32.rs into submodules to satisfy LOC gate
